requirements.txt
streamlit>=1.32.0
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
requests>=2.31.0
